feat: Updated `koa` instrumentation to properly wrap `Router` when using `@koa/router@15.0.0`+
