+++ lib/instrumentation/koa/router-instrumentation.js
@@ -10,7 +10,8 @@ const { MiddlewareSpec, MiddlewareMounterSpec } = require('../../shim/specs')
 module.exports = function instrumentRouter(shim, Router) {
   shim.setFramework(shim.KOA)
 
-  const proto = Router.prototype
+  // @koa/router 15.x exports the Router class as the default export
+  const proto = Router.prototype || Router.default?.prototype
 
   shim.wrapReturn(proto, 'register', wrapMiddleware)
   shim.wrapReturn(proto, 'allowedMethods', wrapAllowedMethods)
