+++ test/versioned/koa/code-level-metrics.test.js
@@ -6,66 +6,21 @@
 'use strict'
 
 const test = require('node:test')
-const http = require('node:http')
 const tspl = require('@matteo.collina/tspl')
-
+const { startServer, makeRequest } = require('./utils')
 const { removeModules } = require('../../lib/cache-buster')
 const helper = require('../../lib/agent_helper')
 const assertClmAttrs = require('../../lib/custom-assertions/assert-clm-attrs')
 
-async function setupApp({ ctx, useKoaRouter, useAtKoaRouter, isCLMEnabled }) {
+async function setupApp({ ctx, isCLMEnabled }) {
   const agent = helper.instrumentMockedAgent({ code_level_metrics: { enabled: isCLMEnabled } })
-  let router
-
-  if (useKoaRouter === true) {
-    const Router = require('koa-router')
-    router = new Router()
-  }
-
-  if (useAtKoaRouter === true) {
-    const Router = require('@koa/router')
-    router = new Router()
-  }
-
   const Koa = require('koa')
   const app = new Koa()
   const server = await startServer(app)
 
   ctx.agent = agent
   ctx.app = app
-  ctx.router = router
   ctx.server = server
-}
-
-async function startServer(app) {
-  return new Promise((resolve) => {
-    const server = app.listen(0, () => resolve(server))
-  })
-}
-
-async function makeRequest(params) {
-  return new Promise((resolve, reject) => {
-    const req = http.request(params, (res) => {
-      if (res.statusCode < 200 || res.statusCode >= 300) {
-        reject(new Error(`Status Code: ${res.statusCode}`))
-        return
-      }
-
-      const data = []
-
-      res.on('data', (chunk) => {
-        data.push(chunk)
-      })
-
-      res.on('end', () => resolve(Buffer.concat(data).toString()))
-    })
-
-    req.on('error', (err) => {
-      reject(err)
-    })
-
-    req.end()
-  })
 }
 
 test('vanilla koa, no router', async (t) => {
@@ -75,7 +30,7 @@ test('vanilla koa, no router', async (t) => {
 
   t.afterEach((ctx) => {
     ctx.nr.server.close()
-    removeModules(['koa', '@koa/router', 'koa-router'])
+    removeModules(['koa'])
     helper.unloadAgent(ctx.nr.agent)
   })
 
@@ -120,148 +75,6 @@ test('vanilla koa, no router', async (t) => {
 
       const response = await makeRequest({ port: server.address().port })
       plan.equal(response, 'done', 'should return the correct data')
-
-      await plan.completed
-    })
-  }
-})
-
-test('using koa-router', async (t) => {
-  t.beforeEach((ctx) => {
-    ctx.nr = {}
-  })
-
-  t.afterEach((ctx) => {
-    ctx.nr.server.close()
-    removeModules(['koa', '@koa/router', 'koa-router'])
-    helper.unloadAgent(ctx.nr.agent)
-  })
-
-  for (const isCLMEnabled of [true, false]) {
-    await t.test(`should ${isCLMEnabled ? 'add' : 'not add'} CLM attributes`, async (t) => {
-      const plan = tspl(t, { plan: 13 })
-
-      await setupApp({ ctx: t.nr, isCLMEnabled, useKoaRouter: true })
-      const { agent, app, router, server } = t.nr
-
-      const Router = require('koa-router')
-      const nestedRouter = new Router()
-
-      nestedRouter.get('/second', function secondMiddleware(ctx) {
-        ctx.body = 'winner winner, chicken dinner'
-      })
-
-      router.use(function appLevelMiddleware(ctx, next) {
-        ctx.body = 'nope, not here'
-        return next()
-      })
-      router.use('/:first', nestedRouter.routes())
-      app.use(router.routes())
-
-      agent.on('transactionFinished', (transaction) => {
-        const [baseSegment] = transaction.trace.getChildren(transaction.trace.root.id)
-        const [dispatch] = transaction.trace.getChildren(baseSegment.id)
-        const [appLevel] = transaction.trace.getChildren(dispatch.id)
-        const [secondMw] = transaction.trace.getChildren(appLevel.id)
-
-        assertClmAttrs(
-          {
-            segments: [
-              {
-                segment: dispatch,
-                name: 'dispatch',
-                filepath: 'koa-router/lib/router.js'
-              },
-              {
-                segment: appLevel,
-                name: 'appLevelMiddleware',
-                filepath: 'code-level-metrics.test.js'
-              },
-              {
-                segment: secondMw,
-                name: 'secondMiddleware',
-                filepath: 'code-level-metrics.test.js'
-              }
-            ],
-            enabled: isCLMEnabled
-          },
-          { assert: plan }
-        )
-      })
-
-      const response = await makeRequest({ port: server.address().port, path: '/123/second' })
-      plan.equal(response, 'winner winner, chicken dinner', 'should return the correct data')
-
-      await plan.completed
-    })
-  }
-})
-
-test('using @koa/router', async (t) => {
-  t.beforeEach((ctx) => {
-    ctx.nr = {}
-  })
-
-  t.afterEach((ctx) => {
-    ctx.nr.server.close()
-    removeModules(['koa', '@koa/router', 'koa-router'])
-    helper.unloadAgent(ctx.nr.agent)
-  })
-
-  for (const isCLMEnabled of [true, false]) {
-    await t.test(`should ${isCLMEnabled ? 'add' : 'not add'} CLM attributes`, async (t) => {
-      const plan = tspl(t, { plan: 13 })
-
-      await setupApp({ ctx: t.nr, isCLMEnabled, useAtKoaRouter: true })
-      const { agent, app, router, server } = t.nr
-
-      const Router = require('@koa/router')
-      const nestedRouter = new Router()
-
-      nestedRouter.get('/second', function secondMiddleware(ctx) {
-        ctx.body = 'winner winner, chicken dinner'
-      })
-
-      router.use(function appLevelMiddleware(ctx, next) {
-        ctx.body = 'nope, not here'
-        return next()
-      })
-      router.use('/:first', nestedRouter.routes())
-      app.use(router.routes())
-
-      agent.on('transactionFinished', (transaction) => {
-        const [baseSegment] = transaction.trace.getChildren(transaction.trace.root.id)
-        const [dispatch] = transaction.trace.getChildren(baseSegment.id)
-        const [appLevel] = transaction.trace.getChildren(dispatch.id)
-        const [secondMw] = transaction.trace.getChildren(appLevel.id)
-
-        assertClmAttrs(
-          {
-            segments: [
-              {
-                segment: dispatch,
-                name: 'dispatch',
-                filepath: '@koa/router/lib/router.js'
-              },
-              {
-                segment: appLevel,
-                name: 'appLevelMiddleware',
-                filepath: 'code-level-metrics.test.js'
-              },
-              {
-                segment: secondMw,
-                name: 'secondMiddleware',
-                filepath: 'code-level-metrics.test.js'
-              }
-            ],
-            enabled: isCLMEnabled
-          },
-          { assert: plan }
-        )
-      })
-
-      const response = await makeRequest({ port: server.address().port, path: '/123/second' })
-      plan.equal(response, 'winner winner, chicken dinner', 'should return the correct data')
 
       await plan.completed
     })
